update_redcap_diff now allows new records
Adds an --allow-new option to update_redcap_diff; this allows the
algorithm to create new records.

This also refactors the test library a bit, to make testing this more
practical.

This also also fixes a deprecation in split_redcap_data, and simplifies
the condense_df method -- previously we were converting the blank values
to NaN and that is not necessary.