Refactor "special REDCap column" handling
update_redcap_diff treats "index" columns (the first column,
redcap_event_name, and the instance columns) specially -- those are the
things that need to match exactly when we're making normal changes, and
are where updates can add to updated_csv with --allow-new

But! We were missing the instance columns in that list.

This changes the script to have a list of INDEX_COLS it will look for,
and if those are present, it'll add them to the index list.

Fixes #2